requirements.txt
colorama
discord.py
stegano
requests
discord_webhook
plyer
selenium
beautifulsoup4
Pillow
PyAutoGUI
pyperclip
aiohttp
pypresence
pyinstaller
dhooks
pystyle
SSL
pyfiglet
fake_useragent
beautifulsoup4
wifipasswords
progress==1.4
progress.bar
subprocess
argon2-cffi
hashlib 
bcrypt 
passlib 
passlib.hash

